Refactor auth and place services to update import paths and enhance API response typing
