+++ client/features/auth/components/RegisterForm.tsx
@@ -8,7 +8,7 @@ import {
 } from "react-native";
 import { useForm, Controller } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
-import { registerSchema } from "@/schemas/authSchema";
+import { registerSchema } from "@/features/auth/schemas/authSchema";
 import { useAuth } from "@/features/auth/context/AuthContext";
 import { Ionicons } from "@expo/vector-icons";
 import { authStyles as styles } from "@/features/auth/styles/auth.styles";
